requirements.txt
starbot-bilibili
redis
graia-ariadne
graia-saya 